backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doce-erp-backend"
version = "0.1.0"
description = "DocERP — Spec-Driven Document Management System for Siliconmotion"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.29",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "celery[redis]>=5.3.6",
    "redis>=5.0.3",
    "langchain>=0.1.16",
    "langchain-community>=0.0.38",
    "langchain-openai>=0.1.3",
    "chromadb>=0.4.24",
    "openpyxl>=3.1.2",
    "gitpython>=3.1.43",
    "python-ldap>=3.4.4",
    "httpx>=0.27.0",
    "pydantic>=2.7.1",
    "pydantic-settings>=2.2.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "reportlab>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "locust>=2.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

# === Ruff (linter + formatter) ===
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
    "ANN", # flake8-annotations
]
ignore = [
    "ANN101",  # missing type annotation for self
    "ANN102",  # missing type annotation for cls
    "S101",    # assert (used in tests)
]
[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "ANN"]
"migrations/**" = ["ANN"]

[tool.ruff.format]
quote-style = "double"

# === mypy (strict) ===
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = ["migrations/", "tests/"]
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
warn_required_dynamic_aliases = true

# === pytest ===
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=70"
